split inventory input handler
