+++ crates/client/src/app/input/mod.rs
@@ -1,1 +1,2 @@
 pub mod hotbar_assignment;
+pub mod inventory;
