+++ crates/client/src/app/input_handler.rs
@@ -2,10 +2,7 @@ use crossterm::event::KeyCode;
 use std::{error::Error, time::Instant};
 
 use lithicrivers_core::{
-    components::{
-        itemkind_name, itemkind_sprite_name, DroppedItem, Inventory as InvComp, ItemKind,
-        ItemStack, Position, SpriteRef,
-    },
+    components::{itemkind_name, Inventory as InvComp, ItemStack},
     game::GameTickResult,
     moves::get_available_moves,
 };
@@ -448,23 +445,8 @@ pub fn handle_input(app: &mut App, key: KeyCode) -> Result<(), Box<dyn Error>> {
             }
         }
     }
-    // Inventory: toggle item auto-pickup
-    if app.ui.current_tab == MenuTab::Inventory
-        && app
-            .ui
-            .keybinds
-            .matches("inventory", "TOGGLE_ITEM_AUTO_PICKUP_KEY", &key)
-    {
-        if let Some(e) = app.core.game.get_player_entity() {
-            if let Ok(mut inv) = app.core.game.world.get::<&mut InvComp>(e) {
-                inv.auto_pickup = !inv.auto_pickup;
-                let state = if inv.auto_pickup { "ON" } else { "OFF" };
-                app.core
-                    .game
-                    .res
-                    .log(format!("Item auto-pickup: {}", state));
-            }
-        }
+    // Handle inventory tab input when active - delegated to input module
+    if crate::app::input::inventory::handle_inventory_input(app, key)? {
         return Ok(());
     }
 
@@ -817,157 +799,6 @@ pub fn handle_input(app: &mut App, key: KeyCode) -> Result<(), Box<dyn Error>> {
                 selected_repair: repair_idx,
                 selected_body_part: part_selection,
             };
-            return Ok(());
-        }
-    }
-
-    // Inventory panel-specific navigation and actions
-    if app.ui.current_tab == MenuTab::Inventory {
-        // Move selection: support Up/Down keys and numpad 8/2 (MOVE_NORTH/SOUTH)
-        if app.ui.keybinds.matches("ui", "CREDITS_SCROLL_UP", &key)
-            || app.ui.keybinds.matches("movement", "MOVE_NORTH", &key)
-        {
-            if let Some(e) = app.core.game.get_player_entity() {
-                if let Ok(inv) = app.core.game.world.get::<&InvComp>(e) {
-                    if !inv.slots.is_empty() {
-                        if app.panels.inventory.selected == 0 {
-                            app.panels.inventory.selected = inv.slots.len() - 1;
-                        } else {
-                            app.panels.inventory.selected -= 1;
-                        }
-                    }
-                }
-            }
-            return Ok(());
-        }
-        if app.ui.keybinds.matches("ui", "CREDITS_SCROLL_DOWN", &key)
-            || app.ui.keybinds.matches("movement", "MOVE_SOUTH", &key)
-        {
-            if let Some(e) = app.core.game.get_player_entity() {
-                if let Ok(inv) = app.core.game.world.get::<&InvComp>(e) {
-                    if !inv.slots.is_empty() {
-                        app.panels.inventory.selected =
-                            (app.panels.inventory.selected + 1) % inv.slots.len();
-                    }
-                }
-            }
-            return Ok(());
-        }
-
-        // Helper to get current selection
-        let mut selected: Option<(ItemKind, u32)> = None;
-        if let Some(e) = app.core.game.get_player_entity() {
-            if let Ok(inv) = app.core.game.world.get::<&InvComp>(e) {
-                if !inv.slots.is_empty() {
-                    let idx = app.panels.inventory.selected.min(inv.slots.len() - 1);
-                    selected = Some((inv.slots[idx].kind, inv.slots[idx].qty));
-                }
-            }
-        }
-
-        // Drop selected item (quantity 1 for now)
-        if app.ui.keybinds.matches("inventory", "DROP_ITEM", &key) {
-            if let Some((kind, qty)) = selected {
-                if qty == 0 {
-                    return Ok(());
-                }
-                if let Some(e) = app.core.game.get_player_entity() {
-                    // Copy player position, then drop immutable borrow before mutating world
-                    let (px, py, pz) = {
-                        let Ok(ppos) = app
-                            .core
-                            .game
-                            .world
-                            .get::<&lithicrivers_core::components::Position>(e)
-                        else {
-                            return Ok(());
-                        };
-                        (ppos.x, ppos.y, ppos.z)
-                    };
-
-                    let drop_qty = 1u32;
-                    // Decrement inventory (mutable borrow scope ends before spawn)
-                    if let Ok(mut inv) = app.core.game.world.get::<&mut InvComp>(e) {
-                        if app.panels.inventory.selected < inv.slots.len() {
-                            let slot = &mut inv.slots[app.panels.inventory.selected];
-                            if slot.qty >= drop_qty {
-                                slot.qty -= drop_qty;
-                                if slot.qty == 0 {
-                                    inv.slots.remove(app.panels.inventory.selected);
-                                    if app.panels.inventory.selected > 0 {
-                                        app.panels.inventory.selected -= 1;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    // Spawn DroppedItem entity with SpriteRef
-                    let _ = app.core.game.world.spawn((
-                        Position {
-                            x: px,
-                            y: py,
-                            z: pz,
-                        },
-                        DroppedItem {
-                            kind,
-                            qty: drop_qty,
-                        },
-                        SpriteRef::new("items", itemkind_sprite_name(kind)),
-                    ));
-                    app.core
-                        .game
-                        .res
-                        .log(format!("Dropped 1 {}", itemkind_name(kind)));
-                }
-            }
-            return Ok(());
-        }
-
-        // Duplicate selected item (cheat)
-        if app
-            .ui
-            .keybinds
-            .matches("inventory", "CHEAT_DUPLICATE_ITEM", &key)
-        {
-            if let Some(e) = app.core.game.get_player_entity() {
-                if let Ok(mut inv) = app.core.game.world.get::<&mut InvComp>(e) {
-                    if !inv.slots.is_empty() {
-                        let idx = app.panels.inventory.selected.min(inv.slots.len() - 1);
-                        let kind = inv.slots[idx].kind;
-                        inv.slots[idx].qty = inv.slots[idx].qty.saturating_add(1);
-                        app.core
-                            .game
-                            .res
-                            .log(format!("Duplicated 1 {}", itemkind_name(kind)));
-                    }
-                }
-            }
-            return Ok(());
-        }
-
-        // Destroy selected item (remove 1)
-        if app.ui.keybinds.matches("inventory", "DESTROY_ITEM", &key) {
-            if let Some(e) = app.core.game.get_player_entity() {
-                if let Ok(mut inv) = app.core.game.world.get::<&mut InvComp>(e) {
-                    if !inv.slots.is_empty() {
-                        let idx = app.panels.inventory.selected.min(inv.slots.len() - 1);
-                        let kind = inv.slots[idx].kind;
-                        if inv.slots[idx].qty > 0 {
-                            inv.slots[idx].qty -= 1;
-                            if inv.slots[idx].qty == 0 {
-                                inv.slots.remove(idx);
-                                if app.panels.inventory.selected > 0 {
-                                    app.panels.inventory.selected -= 1;
-                                }
-                            }
-                            app.core
-                                .game
-                                .res
-                                .log(format!("Destroyed 1 {}", itemkind_name(kind)));
-                        }
-                    }
-                }
-            }
             return Ok(());
         }
     }
